decomp overlay_80_02233688 -> frontier_command_battle_arcade, 1

+++ include/overlay_80.h
@@ -2,7 +2,7 @@
 #define POKEHEARTGOLD_OVERLAY_80_H
 
 BOOL BattleArcade_MultiplayerCheck(int type);
-int ov80_022384BC(u8 a0);
+u8 ov80_022384BC(u8 a0);
 u8 BattleArcade_GetMonCount(u8 type, int a1);
 u8 BattleArcade_GetOpponentMonCount(u8 type, int a1);
